+++ verifier.go
@@ -127,7 +127,10 @@ func (v *verifier) verifySearchTree() (map[uint]bool, error) {
 }
 
 func (v *verifier) verifyDataSectionSeparator() error {
-	separatorStart := v.reader.Metadata.NodeCount * v.reader.Metadata.RecordSize / 4
+	separatorStart := searchTreeSizeBytes(
+		v.reader.Metadata.NodeCount,
+		v.reader.Metadata.RecordSize,
+	)
 	separatorEnd := separatorStart + dataSectionSeparatorSize
 	if separatorEnd < separatorStart || separatorEnd > uint(len(v.reader.buffer)) {
 		return mmdberrors.NewInvalidDatabaseError(